smartShop/requirements/local.txt
South
bpython
django-crispy-forms
django-tastypie==0.9.16
python-mimeparse    #tastypie dependencies-
python-dateutil     #tastypie dependencies
